fix(common): api swagger docs
