+++ common/package.json
@@ -20,6 +20,9 @@
     "mocha-junit-reporter": "^2.0.2"
   },
   "dependencies": {
+    "express": "^4.17.1",
+    "swagger-jsdoc": "^6.1.0",
+    "swagger-ui-express": "^4.1.6",
     "uuid": "^8.3.2",
     "validate.js": "^0.13.1"
   }
